simplified scripts with args and added basic docker cmds
